+++ VergiNoDogrula.WPF/Services/IBackupService.cs
@@ -1,3 +1,5 @@
+using VergiNoDogrula.WPF.Models;
+
 namespace VergiNoDogrula.WPF.Services;
 
 /// <summary>
@@ -23,4 +25,17 @@ internal interface IBackupService
     /// </summary>
     /// <returns>True if backup should be performed, false otherwise.</returns>
     bool IsBackupDue();
+
+    /// <summary>
+    /// Gets the list of backup files in the backup folder.
+    /// </summary>
+    /// <returns>List of backup file information ordered by creation time (newest first).</returns>
+    Task<List<BackupFileInfo>> GetBackupFilesAsync();
+
+    /// <summary>
+    /// Checks if any backup files exist that are newer than the specified UTC time.
+    /// </summary>
+    /// <param name="lastBackupTimeUtc">The UTC timestamp to compare against backup file creation times.</param>
+    /// <returns>True if one or more backup files were created after the specified time; otherwise, false.</returns>
+    Task<bool> HasNewerBackupThanAsync(DateTime lastBackupTimeUtc);
 }