+++ VergiNoDogrula.WPF/ViewModels/TaxPayerCollectionVM.cs
@@ -29,9 +29,11 @@ namespace VergiNoDogrula.WPF.ViewModels
             DeleteTaxPayerCommand = new DeleteTaxPayerCommand();
             SaveTaxPayerCommand = new SaveTaxPayerCommand();
             BackupDatabaseCommand = new BackupDatabaseCommand();
+            ShowBackupListCommand = new ShowBackupListCommand();
             ShowAboutCommand = new ShowAboutCommand();
         }
 
+        public IBackupService BackupService => _backupService;
 
         public string Status
         {
@@ -50,6 +52,7 @@ namespace VergiNoDogrula.WPF.ViewModels
         public ICommand EmptySearchStringCommand { get; }
         public ICommand SaveTaxPayerCommand { get; }
         public ICommand BackupDatabaseCommand { get; }
+        public ICommand ShowBackupListCommand { get; }
         public ICommand ShowAboutCommand { get; }
 
         public bool IsSearching { get; private set; }